requirements.txt
Django==1.8.1
django-ajax-selects==1.3.6
